chaged player script, healthregen and spikes' script
player script was divided in various module scripts, with the class Player being the cornerstone.
The player will no longer check if it's colliding with healthregen or dangerous agents. each one of them will prompt the player to heal/take damage (they are the ones checking if the player is colliding, and not the other way around)